+++ profiles/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
-from profiles.models.generic import Email, Profile
+from profiles.models.combined import Profile
+from profiles.models.generic import Email
 from profiles.models.staff_sso import StaffSSOProfile, StaffSSOProfileEmail
