+++ profiles/services/__init__.py
@@ -1,6 +1,14 @@
 # This is the entrypoint service that coordinates between the sub-services
 # If in doubt about what to use, you should probably be using this
 
+from profiles.models.combined import Profile
+
+from .combined import get_by_id as get_combined_by_id
+
+
+def get_by_id(sso_email_id: str) -> Profile:
+    return get_combined_by_id(sso_email_id=sso_email_id)
+
 
 def generate_combined_profile_fields(sso_user_id: str):
     """
@@ -16,7 +24,7 @@ def create_from_sso(
     first_name: str,
     last_name: str,
     emails: list[dict],
-):
+) -> Profile:
     """A central function to create all relevant profile details"""
     # staff_sso.create()
     # payload = generate_combined_profile_values()
